Some small changes  for workflow
I have done small correction, where I found things unclear. It may be good  to use a real 'mdp' (linked to a specific force field)  to avoid to confuse the reader or clearly state that what is there can not use as it is. I will also be good to add link to BIOEXCEL webinar (e.i. mdp parameter webinar). 
For the future  it will be useful to integrate the tools part by taking an idea what is unclear for users (based on the questions/answer on the forum). Those information will feel the gap that the user notice between the tool description in the manual and the use of the tools.